// src/components/common/context-menu/types.ts
import type { JSX } from 'solid-js';

export interface ContextMenuItem {
  label: string | (() => string);
  icon?: (props: { class?: string }) => JSX.Element;
  onClick: () => void;
  disabled?: boolean | (() => boolean);
  closeOnSelect?: boolean;
  class?: string;
  showWhen?: () => boolean;
  children?: ContextMenuItem[];
}